+++ core/src/commonMain/kotlin/com/powersync/db/driver/ReadPool.kt
@@ -1,7 +1,8 @@
-package com.powersync.db.internal
+package com.powersync.db.driver
 
+import androidx.sqlite.SQLiteConnection
+import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PowerSyncException
-import com.powersync.PsSqlDriver
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
@@ -11,16 +12,20 @@ import kotlinx.coroutines.channels.ClosedSendChannelException
 import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 
-internal class ConnectionPool(
-    factory: () -> PsSqlDriver,
+/**
+ * The read-part of a [SQLiteConnectionPool] backed by connections owned by the PowerSync SDK.
+ */
+@OptIn(ExperimentalPowerSyncAPI::class)
+internal class ReadPool(
+    factory: () -> SQLiteConnection,
     size: Int = 5,
     private val scope: CoroutineScope,
 ) {
-    private val available = Channel<Pair<TransactorDriver, CompletableDeferred<Unit>>>()
+    private val available = Channel<Pair<SQLiteConnection, CompletableDeferred<Unit>>>()
     private val connections: List<Job> =
         List(size) {
             scope.launch {
-                val driver = TransactorDriver(factory())
+                val driver = factory()
                 try {
                     while (true) {
                         val done = CompletableDeferred<Unit>()
@@ -33,31 +38,30 @@ internal class ConnectionPool(
                         done.await()
                     }
                 } finally {
-                    driver.driver.close()
+                    driver.close()
                 }
             }
         }
 
-    suspend fun <R> withConnection(action: suspend (connection: TransactorDriver) -> R): R {
-        val (connection, done) =
-            try {
-                available.receive()
-            } catch (e: PoolClosedException) {
-                throw PowerSyncException(
-                    message = "Cannot process connection pool request",
-                    cause = e,
-                )
-            }
+    suspend fun <T> read(block: suspend (SQLiteConnectionLease) -> T): T {
+        val (connection, done) = try {
+            available.receive()
+        } catch (e: PoolClosedException) {
+            throw PowerSyncException(
+                message = "Cannot process connection pool request",
+                cause = e,
+            )
+        }
 
         try {
-            return action(connection)
+            return block(RawConnectionLease(connection))
         } finally {
             done.complete(Unit)
         }
     }
 
-    suspend fun <R> withAllConnections(action: suspend (connections: List<TransactorDriver>) -> R): R {
-        val obtainedConnections = mutableListOf<Pair<TransactorDriver, CompletableDeferred<Unit>>>()
+    suspend fun <R> withAllConnections(action: suspend (connections: List<SQLiteConnection>) -> R): R {
+        val obtainedConnections = mutableListOf<Pair<SQLiteConnection, CompletableDeferred<Unit>>>()
 
         try {
             /**
@@ -87,6 +91,7 @@ internal class ConnectionPool(
         available.cancel(PoolClosedException)
         connections.joinAll()
     }
+
 }
 
 internal object PoolClosedException : CancellationException("Pool is closed")