+++ core/src/watchosMain/kotlin/com/powersync/DatabaseDriverFactory.watchos.kt
@@ -1,9 +1,20 @@
 package com.powersync
 
-import co.touchlab.sqliter.DatabaseConnection
+import androidx.sqlite.driver.bundled.BundledSQLiteDriver
 import com.powersync.static.powersync_init_static
 
-internal actual fun DatabaseConnection.loadPowerSyncSqliteCoreExtension() {
+@Suppress(names = ["EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING"])
+public actual class DatabaseDriverFactory {
+    internal actual fun resolveDefaultDatabasePath(dbFilename: String): String {
+        return appleDefaultDatabasePath(dbFilename)
+    }
+}
+
+public actual fun BundledSQLiteDriver.addPowerSyncExtension() {
+    didLoadExtension
+}
+
+private val didLoadExtension by lazy {
     val rc = powersync_init_static()
     if (rc != 0) {
         throw PowerSyncException(
@@ -14,4 +25,6 @@ internal actual fun DatabaseConnection.loadPowerSyncSqliteCoreExtension() {
                 ),
         )
     }
+
+    true
 }