+++ static-sqlite-driver/src/commonMain/kotlin/com/powersync/sqlite3/StaticSqliteDriver.kt
@@ -1,9 +1,0 @@
-package com.powersync.sqlite3
-
-/**
- * An empty Kotlin object.
- *
- * This package needs to provide a source to be published correctly. The only purpose of this package is to provide
- * build scripts linking SQLite statically however, so this empty object is defined for publishing only.
- */
-public object StaticSqliteDriver